Add TItem and TCollection BindableListView.
