+++ src/UnityMvvmToolkit.UnityPackage/Assets/Plugins/UnityMvvmToolkit/Runtime/UITK/BindableUIElements/BindableListView.TItem.TCollection.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c56f30f3cb5b45f196d47ee4e73e00d5
+timeCreated: 1688698861
