+++ src/UnityMvvmToolkit.UnityPackage/Assets/Plugins/UnityMvvmToolkit/Runtime/UITK/BindableUIElements/Uxmls/BindableListView.TItem.TCollection.Uxml.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 3da29e416dc6494995a8366ff9df23e2
+timeCreated: 1688699627
